Fix permissions of logs directory
The aloja-web/logs directory was mounted as 755 vagrant:vagrant, but PHP scripts are executed as www-data:www-data, which can't write to the log file and aborts the request.

Now the folder is mounted as 775 www-data:vagrant, so both users can modify the file (no more errors!).

Also, the file aloja_dev.log has been removed, because it was gitignore'd but still present in the repo, so changes were still tracked.