Benchmark configuration refactor.
A new function in common_benchmarks permits configuring any benchmark
in a centralized way without duplicating code. It uses specific variables
"use_$engine{hadoop,hive,spark}" to set the requirements to download,
prepare and start necessary services.

The objective of such modification is to reduce benchmark definition
complexity, code duplication, code than runs unnecessarily more than once
and readability.

Some sourcing is still maintained to provide support for legacy benchmarks
until they are all ported to the new system.

Modified benchmark configuration conditionals to avoid incorrect uses.

BugFix: Hive and Derby only benchmarks such as hive-test and derby-test
that are not yet ported to the nev configuration system need to source
Hadoop and Java for them to work. Eventually once all benchmarks are
ported, this sources need to go away.

Added a init method in each engine that handles the initilization and
configuration.

Also, initilization of monitoring tools is moved to
benchmark_suite_start_config as part of the benchmark configuration, this
avoids an issue where some monitor tools where not correctly started.

Modified the way we check if an engine is needed, using use_$engine=1.
This adds security to not initilize the engine if some garbage string is
set to the variables.

Solve package download and configuration sync